feat(ui-portal): remove `SubtreePortal` private component, not needed with react 16
Closes: INSTUI-3071

Since version 16, React supports Portals. InstUI doesn't support React 15 and below, so this
workaround could be deleted, and `Portal` component was simplified to use the native React portals.
