+++ packages/ui-portal/src/Portal/index.tsx
@@ -28,13 +28,13 @@ import ReactDOM from 'react-dom'
 
 import { element } from '@instructure/ui-prop-types'
 
-import { ReactPortal } from './ReactPortal'
-import { SubtreePortal } from './SubtreePortal'
-import { BidirectionalProps } from '@instructure/ui-i18n'
+import { passthroughProps } from '@instructure/ui-react-utils'
+import { bidirectional, BidirectionalProps } from '@instructure/ui-i18n'
 
-import { PortalNode } from './PortalType'
-
-const IS_CREATE_PORTAL_SUPPORTED = typeof ReactDOM.createPortal === 'function'
+/**
+ * The DOM Node of the Portal. It is created as a `span` element.
+ */
+type PortalNode = HTMLSpanElement | null
 
 type Props = {
   open?: boolean
@@ -46,13 +46,18 @@ type Props = {
 } & HTMLAttributes<Props> &
   BidirectionalProps
 
+type State = {
+  mountNode: Element
+}
+
 /**
 ---
 category: components/utilities
 ---
 @module Portal
 **/
-class Portal extends Component<Props> {
+@bidirectional()
+class Portal extends Component<Props, State> {
   static propTypes = {
     /**
      * Wheter or not the `<Portal />` is open
@@ -60,7 +65,7 @@ class Portal extends Component<Props> {
     open: PropTypes.bool,
 
     /**
-     * Callback fired when `<Portal />` content has been mounted in the DOM
+     * Callback fired when `<Portal />` content has been mounted in the DOM. Ha the Portal DOMNode as parameter.
      */
     onOpen: PropTypes.func,
 
@@ -84,9 +89,11 @@ class Portal extends Component<Props> {
      */
     children: PropTypes.node,
     /**
-     * provides a reference to the underlying html element
+     * Provides a reference to the underlying html element. Ha the Portal DOMNode as parameter.
      */
-    elementRef: PropTypes.func
+    elementRef: PropTypes.func,
+    // eslint-disable-next-line react/require-default-props
+    dir: PropTypes.oneOf(Object.values(bidirectional.DIRECTION))
   }
 
   static defaultProps = {
@@ -99,23 +106,147 @@ class Portal extends Component<Props> {
     elementRef: () => {}
   }
 
+  constructor(props: Props) {
+    super(props)
+
+    this.state = {
+      mountNode: this.findMountNode(props)
+    }
+  }
+
   DOMNode: PortalNode = null
 
-  handleElementRef = (el: PortalNode) => {
-    if (el) {
-      this.DOMNode = el
+  componentDidMount() {
+    // If Portal is mounting in an open condition fire onOpen handler
+    if (this.props.open && typeof this.props.onOpen === 'function') {
+      this.props.onOpen(this.DOMNode)
+    }
+  }
+
+  componentDidUpdate(prevProps: Props) {
+    const mountNode = this.findMountNode(this.props)
+
+    if (mountNode !== this.state.mountNode) {
+      // set state here to make the component re-render
+      this.setState({ mountNode })
+    }
+
+    // If Portal was closed but is now open fire onOpen handler
+    if (
+      this.props.open &&
+      !prevProps.open &&
+      typeof this.props.onOpen === 'function'
+    ) {
+      this.props.onOpen(this.DOMNode)
+    }
+
+    // If Portal was open but is now closed fire onClose handler
+    if (
+      !this.props.open &&
+      prevProps.open &&
+      typeof this.props.onClose === 'function'
+    ) {
+      this.props.onClose()
+    }
+  }
+
+  componentWillUnmount() {
+    this.removeNode()
+
+    // If Portal was open fire onClose handler
+    if (this.props.open && typeof this.props.onClose === 'function') {
+      this.props.onClose()
+    }
+  }
+
+  removeNode() {
+    if (
+      this.DOMNode &&
+      this.DOMNode.parentNode &&
+      typeof this.DOMNode.parentNode.removeChild === 'function'
+    ) {
+      this.DOMNode.parentNode.removeChild(this.DOMNode)
+      this.DOMNode = null
+
       if (typeof this.props.elementRef === 'function') {
-        this.props.elementRef(el)
+        this.props.elementRef(this.DOMNode)
       }
     }
   }
 
-  render() {
-    return IS_CREATE_PORTAL_SUPPORTED ? (
-      <ReactPortal {...this.props} elementRef={this.handleElementRef} />
-    ) : (
-      <SubtreePortal {...this.props} elementRef={this.handleElementRef} />
-    )
+  insertNode() {
+    const {
+      open,
+      insertAt,
+      onOpen,
+      onClose,
+      mountNode,
+      children,
+      elementRef,
+      dir,
+      ...props
+    } = this.props
+
+    // Create node if it doesn't already exist
+    if (!this.DOMNode) {
+      const node = document.createElement('span')
+      const attributes: Partial<Props> = {
+        ...passthroughProps(props as Partial<Props>),
+        dir
+      }
+
+      ;(Object.keys(attributes) as Array<keyof Props>).forEach((name) => {
+        node.setAttribute(name, attributes[name] as string)
+      })
+
+      if (typeof elementRef === 'function') {
+        elementRef(node)
+      }
+
+      this.DOMNode = node
+    }
+
+    // Append node to container if it isn't already
+    if (this.DOMNode.parentNode !== this.state.mountNode) {
+      if (insertAt === 'bottom') {
+        this.state.mountNode.appendChild(this.DOMNode)
+      } else {
+        this.state.mountNode.insertBefore(
+          this.DOMNode,
+          this.state.mountNode.firstChild
+        )
+      }
+    }
+
+    return this.DOMNode
+  }
+
+  findMountNode(props: Props) {
+    let mountNode: Element | undefined
+
+    if (typeof props.mountNode === 'function') {
+      mountNode = props.mountNode()
+    } else if (props.mountNode) {
+      mountNode = props.mountNode
+    }
+
+    if (!mountNode || !mountNode.nodeName) {
+      mountNode = document.body
+    }
+
+    return mountNode
+  }
+
+  get node() {
+    return this.DOMNode
+  }
+
+  render(): React.ReactPortal | null {
+    const { children } = this.props
+
+    return this.props.open && React.Children.count(children) > 0
+      ? ReactDOM.createPortal(children, this.insertNode())
+      : null
   }
 }
